refactor sherloModule
