--- packages/react-native-storybook/src/nativeModule/index.tsx
@@ -1,1 +1,0 @@
-export { default as getSherloModule } from './getSherloModule';
