+++ packages/react-native-storybook/src/runnerBridge/actions/create.ts
@@ -1,7 +1,5 @@
-import { getSherloModule } from '../../nativeModule';
+import sherloModule from '../../sherloModule';
 import { LogFn } from '../types';
-
-const sherloModule = getSherloModule();
 
 const create = (path: string, log: LogFn) => async (): Promise<void> => {
   await sherloModule.mkdir(path).catch(() => {
