+++ packages/react-native-storybook/src/runnerBridge/actions/getConfig.ts
@@ -1,9 +1,9 @@
-import { getSherloModule } from '../../nativeModule';
+import sherloModule from '../../sherloModule';
 import { getGlobalStates, isObject } from '../../utils';
 import { LogFn } from '../types';
 import { verifySignature } from '../utils';
 
-const sherloModule = getSherloModule();
+const { readFile } = sherloModule;
 
 // TODO: change name to getDeviceConfig? + return type fix
 const getConfig = (path: string, log?: LogFn) => async (): Promise<any> => {
@@ -12,8 +12,8 @@ const getConfig = (path: string, log?: LogFn) => async (): Promise<any> => {
   }
 
   // TODO: add comment to explain replaceAll code
-  const configString = (await sherloModule.readFile(path)).replaceAll('\n', '');
-  const signature = (await sherloModule.readFile(`${path}.sig`)).replaceAll('\n', '');
+  const configString = (await readFile(path)).replaceAll('\n', '');
+  const signature = (await readFile(`${path}.sig`)).replaceAll('\n', '');
 
   const isSignatureValid = verifySignature(configString, signature);
   if (!isSignatureValid) {
