+++ packages/react-native-storybook/src/runnerBridge/actions/getState.ts
@@ -1,15 +1,15 @@
-import { getSherloModule } from '../../nativeModule';
+import sherloModule from '../../sherloModule';
 import { getGlobalStates } from '../../utils';
 import { LogFn, RunnerState } from '../types';
 
-const sherloModule = getSherloModule();
+const { readFile } = sherloModule;
 
 const getState = (path: string, log: LogFn) => async (): Promise<RunnerState> => {
   if (getGlobalStates().isIntegrationTest) {
     throw new Error("We're not using cached state in render test mode");
   }
 
-  const stateFile = await sherloModule.readFile(path);
+  const stateFile = await readFile(path);
   const state = JSON.parse(stateFile) as RunnerState;
 
   log('getState:response', { state });
