+++ packages/react-native-storybook/src/sherloModule/index.tsx
@@ -1,0 +1,1 @@
+export { default } from './sherloModule';
