+++ packages/react-native-storybook/src/sherloModule/utils/normalizeFilePath.ts
